fix warnings

+++ src/xtd_forms_native_wxwidgets/src/xtd/forms/native/wxwidgets/image_list.cpp
@@ -23,11 +23,11 @@ void image_list::destroy(intptr_t image_list) {
 
 void image_list::insert_item(intptr_t image_list, size_t pos, const drawing::image& image) {
   if (image_list == 0) return;
-  if (pos == reinterpret_cast<wxImageList*>(image_list)->GetImageCount())
+  if (pos == static_cast<size_t>(reinterpret_cast<wxImageList*>(image_list)->GetImageCount()))
     reinterpret_cast<wxImageList*>(image_list)->Add(*reinterpret_cast<wxImage*>(image.handle()));
   else {
     std::list<wxBitmap> bitmaps;
-    for (int index = reinterpret_cast<wxImageList*>(image_list)->GetImageCount() - 1; index > pos; index--) {
+    for (size_t index = static_cast<size_t>(reinterpret_cast<wxImageList*>(image_list)->GetImageCount()) - 1; index > pos; index--) {
       bitmaps.push_front(reinterpret_cast<wxImageList*>(image_list)->GetBitmap(index));
       reinterpret_cast<wxImageList*>(image_list)->Remove(index);
     }
